// src/lib/interfaces/env.ts
export interface Env {
  heliusApiKey: string | undefined;
  protocolProgramIdPubkey: string | undefined;
  nonceAuthorityAccountSecretKey: string | undefined;
  nonceAuthorityAccountPubkey: string | undefined;
  nonceAccountPubkey: string | undefined;
  nonceAccountSecretKey: string | undefined;
}